Fix Transmission piece bitmap creation
The Marshal() function in BitArray returns a serialized representation of the
BitArray object itself, which consists of much more than just the actual bit
content : it starts with an identifier and two 64-bit integers.
This caused the returned bit array to always contain some zeroes in the
beginning, which were reported by GUI clients as missing pieces despite, for
example, the torrent being completed.
This commit replaces the bitarray usage with manually creating the bitmap.